### 2024-01-06 #6182 @editorialbot
[PRE REVIEW]: OpenRepGrid: An R Package for the Analysis of Repertory Grid Data
**Submitting author:** @markheckmann (<a href="http://orcid.org/0000-0002-0736-7417">Mark Heckmann</a>)
**Repository:** https://github.com/markheckmann/OpenRepGrid
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.1.15
