### 2024-01-07 039f5e8
add JOSS badge to README [no ci]
