+++ README.Rmd
@@ -25,6 +25,7 @@ library(OpenRepGrid)
 [![AppVeyor build status](https://ci.appveyor.com/api/projects/status/github/markheckmann/OpenRepGrid?branch=master&svg=true)](https://ci.appveyor.com/project/markheckmann/OpenRepGrid)
 [![Codecov test coverage](https://codecov.io/gh/markheckmann/OpenRepGrid/branch/master/graph/badge.svg)](https://codecov.io/gh/markheckmann/OpenRepGrid?branch=master)
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![status](https://joss.theoj.org/papers/88b1ba29095623f416d9fe944afe5cd8/status.svg)](https://joss.theoj.org/papers/88b1ba29095623f416d9fe944afe5cd8)
 <!-- badges: end -->
 
 <br>
