### 2024-01-07 #6182 @markheckmann
Dear Editors,

the following two reviewers would be a good fit. They were already reviewers on this [JOSS submission last year](https://github.com/openjournals/joss-reviews/issues/3292). Hence, they are already familiar with the repertory grid technique.

- [brunaw](https://github.com/brunaw)
- [aj2duncan](https://github.com/aj2duncan)
