### 2024-01-14 2b2183c
add total downloads badge [no ci]
