+++ DESCRIPTION
@@ -17,7 +17,7 @@ Description: Analyze repertory grids, a qualitative-quantitative
     2004, ISBN: 978-0-470-09080-0). The package is part of the The package is part of the 
     <https://openrepgrid.org/> project.
 Version: 0.1.15
-Date: 2023-12-26
+Date: 2024-02-18
 Encoding: UTF-8
 URL: https://github.com/markheckmann/OpenRepGrid, http://markheckmann.github.io/OpenRepGrid/
 Imports:
@@ -70,7 +70,7 @@ Collate:
     'utils-import.r'
     'utils.r'
     'zzz.r'
-RoxygenNote: 7.2.3
+RoxygenNote: 7.3.0
 NeedsCompilation: no
 Suggests: 
     testthat (>= 2.1.0),
