### 2024-02-18 5c63b63
fix addiotional CRAN comments
fix docs

fix more check comments

some more fixes

fix docs

more fixes

more fixes

add comma
