+++ dev/200104_leave_one_out_estimation.R
@@ -4,24 +4,24 @@
 
 
 #' Resample constructs
-#' 
-#' The goal of resampling is to build variations of a single grid. 
-#' Two variants are implemented: The first is the \emph{leave-n-out} approach which 
-#' builds all possible grids when dropping n constructs. The second is a 
+#'
+#' The goal of resampling is to build variations of a single grid.
+#' Two variants are implemented: The first is the \emph{leave-n-out} approach which
+#' builds all possible grids when dropping n constructs. The second is a
 #' \emph{bootstrap} approach, randomly drawing n constructs from the grid.
-#' 
+#'
 #' @param n Number of constructs to drop or to sample in each generated grid. Defaults
-#'   to 1 for \code{grids_leave_n_out} and to the number of constructs for
-#'   \code{grids_bootstrap}.
+#'   to 1 for `grids_leave_n_out` and to the number of constructs for
+#'   `grids_bootstrap`.
 #' @return List of grids.
 #' @export
 #' @rdname resampling
-#' 
+#'
 grids_leave_n_out <- function(x, n = 1)
 {
   nc <- getNoOfConstructs(x)    # size construct system
   prop <- n / nc        # proportion left out
-  if (prop > .4) 
+  if (prop > .4)
     warning("Be aware that you leave more than 40% or more of the constructs", call. = FALSE)
   n.subset <- nc - n
   l <- combn(seq_len(nc), n.subset, simplify = F) # list of subset indexes
@@ -34,12 +34,12 @@ grids_leave_n_out <- function(x, n = 1)
 #' @param replace Resample constructs with replacement?
 #' @export
 #' @rdname resampling
-#' 
+#'
 grids_bootstrap <- function(x, n = nrow(x), reps = 100, replace = TRUE)
 {
   nc <- getNoOfConstructs(x)    # size construct system
   prop <- n / nc                # proportion to sample
-  if (prop < .6) 
+  if (prop < .6)
     warning("Be aware that you resample less than 60% of the constructs", call. = FALSE)
   l_i <- replicate(reps, sample(seq_len(nc), n, replace = replace), simplify = FALSE)
   l <- lapply(l_i, function(i, x) x[i, ], x = x)
@@ -48,7 +48,7 @@ grids_bootstrap <- function(x, n = nrow(x), reps = 100, replace = TRUE)
 
 
 #' Number of constructs and elements for grids in repgridlist
-#' @return Numeric vector with the number of constructs or elements 
+#' @return Numeric vector with the number of constructs or elements
 #'   of each grid in repgridlist object.
 #' @export
 #' @keywords internal
