+++ pkgdown.yml
@@ -25,7 +25,7 @@ articles:
   settings: web/settings.html
   visualization-bertin: web/visualization-bertin.html
   visualization-biplot: web/visualization-biplot.html
-last_built: 2024-02-18T17:19Z
+last_built: 2024-02-18T19:19Z
 urls:
   reference: https://docs.openrepgrid.org/reference
   article: https://docs.openrepgrid.org/articles
