+++ .aspell/defaults.R
@@ -1,0 +1,5 @@
+# see: https://ropensci.org/blog/2022/01/21/ropensci-news-digest-january-2022/#to-quote-or-not-to-quote-non-existing-words-in-description
+Rd_files <- vignettes <- R_files <- description <-
+  list(encoding = "UTF-8",
+       language = "en",
+       dictionaries = c("en_stats", "openrepgrid"))
