+++ DESCRIPTION
@@ -13,11 +13,11 @@ Authors@R: c(
 Description: Analyze repertory grids, a qualitative-quantitative 
     data collection technique devised by George A. Kelly in the 1950s. Today, grids are used across 
     various domains ranging from clinical psychology to marketing. The package contains functions 
-    to quantitatively analyze and visualize repertory grid data (e.g. Fransella, Bell, & Bannister, 
+    to quantitatively analyze and visualize repertory grid data (e.g. 'Fransella', 'Bell', & 'Bannister', 
     2004, ISBN: 978-0-470-09080-0). The package is part of the The package is part of the 
     <https://openrepgrid.org/> project.
 Version: 0.1.15
-Date: 2024-02-18
+Date: 2024-02-19
 Encoding: UTF-8
 URL: https://github.com/markheckmann/OpenRepGrid
 Imports:
