### 2024-03-01 #6182 @Matherion
God, sorry - I'm never on GitHub any more. I'm trying to basically avoid closed source software when I can, so I first went to GitLab ad now moved to CodeBerg, so GitHub is quite 'ancient history' for me by now, sorry!!!

This looks like a great paper, and I'd be willing to review it if that's still an option as far as you're concerned.

However 😕 We're in the process of moving house, and the whole selling/buying thing, apparently, is quite exhausting and stressful. This means I expect bandwidth to only increase enough to do this review towards the end of April, early May...

I know this is probably not great news, but I thought I'd offer it nonetheless. Better than ghosting you 😬 I'm glad I happened to see the GitHub email (which also goes to an account I don't use too much any more 😬). But sorry for having been nonresponsive for basically the entire year! 🙂

If you're responding and I'm not responding sufficiently timely, I'm also at @matherion@mastodon.nl!

Have a good weekend! 🙂
