### 2024-04-13 5f5a35e
`rgl` installation is now optional (close #41)
