+++ R/rgl-3d.r
@@ -12,18 +12,18 @@
 #' @keywords internal
 rglDrawStandardAxes <- function(max.dim = 1, lwd = 1, a.cex = 1.1, a.col = "black",
                                 a.radius = .05, labels = TRUE, spheres = FALSE, ...) {
-  lines3d(c(0, max.dim), c(0, 0), c(0, 0), lwd = lwd, col = a.col)
-  lines3d(c(0, 0), c(0, max.dim), c(0, 0), lwd = lwd, col = a.col)
-  lines3d(c(0, 0), c(0, 0), c(0, max.dim), lwd = lwd, col = a.col)
+  rgl::lines3d(c(0, max.dim), c(0, 0), c(0, 0), lwd = lwd, col = a.col)
+  rgl::lines3d(c(0, 0), c(0, max.dim), c(0, 0), lwd = lwd, col = a.col)
+  rgl::lines3d(c(0, 0), c(0, 0), c(0, max.dim), lwd = lwd, col = a.col)
   if (labels) {
-    text3d(max.dim, 0, 0, "X", cex = a.cex, adj = c(1, 1), col = a.col)
-    text3d(0, max.dim, 0, "Y", cex = a.cex, adj = c(1, 1), col = a.col)
-    text3d(0, 0, max.dim, "Z", cex = a.cex, adj = c(1, 1), col = a.col)
+    rgl::text3d(max.dim, 0, 0, "X", cex = a.cex, adj = c(1, 1), col = a.col)
+    rgl::text3d(0, max.dim, 0, "Y", cex = a.cex, adj = c(1, 1), col = a.col)
+    rgl::text3d(0, 0, max.dim, "Z", cex = a.cex, adj = c(1, 1), col = a.col)
   }
   if (spheres) {
-    spheres3d(max.dim, 0, 0, radius = a.radius, col = a.col)
-    spheres3d(0, max.dim, 0, radius = a.radius, col = a.col)
-    spheres3d(0, 0, max.dim, radius = a.radius, col = a.col)
+    rgl::spheres3d(max.dim, 0, 0, radius = a.radius, col = a.col)
+    rgl::spheres3d(0, max.dim, 0, radius = a.radius, col = a.col)
+    rgl::spheres3d(0, 0, max.dim, radius = a.radius, col = a.col)
   }
 }
 # open3d()
@@ -42,26 +42,32 @@ rglDrawStandardAxes <- function(max.dim = 1, lwd = 1, a.cex = 1.1, a.col = "blac
 rglDrawStandardEllipses <- function(max.dim = 1, lwd = 1, col = "black") {
   x <- seq(0, 2 * pi, len = 361)
   x <- data.frame(sin(x), cos(x)) * max.dim
-  lines3d(x[, 1], x[, 2], 0, col = col, lwd = lwd)
-  lines3d(x[, 1], 0, x[, 2], col = col, lwd = lwd)
-  lines3d(0, x[, 1], x[, 2], col = col, lwd = lwd)
+  rgl::lines3d(x[, 1], x[, 2], 0, col = col, lwd = lwd)
+  rgl::lines3d(x[, 1], 0, x[, 2], col = col, lwd = lwd)
+  rgl::lines3d(0, x[, 1], x[, 2], col = col, lwd = lwd)
 }
 
 
 
 rglDrawElementPoints <- function(coords, dim = 1:3, e.radius = .1, e.sphere.col = "black", ...) {
+  if (!requireNamespace("rgl", quietly = TRUE)) {
+    stop("The 'rgl' package is required to use OpenRepGrid's 3D features => please install 'rgl'.", call. = FALSE)
+  }
   coords <- coords[, dim]
[…]
