+++ README.Rmd
@@ -7,7 +7,7 @@ output: github_document
 ```{r setup, include=FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
-  echo=FALSE,
+  echo = FALSE,
   comment = "#",
   fig.path = "man/figures/"
 )
