+++ NAMESPACE
@@ -207,7 +207,6 @@ import(methods)
 import(plyr, except = c(failwith,id,count,mutate,desc,rename,summarize,summarise,filter,arrange))
 import(psych)
 import(pvclust)
-import(rgl)
 import(stats, except=c(lag,filter))
 import(stringr)
 import(utils)
