### 2024-06-19 #6182 @arfon
:wave: @yiqunchen @stemangiola @rlbarter @AliciaSchep would any of you be willing to review this submission for JOSS? The submission under consideration is _OpenRepGrid: An R Package for the Analysis of Repertory Grid Data_

Many thanks
Arfon
