### 2024-06-21 #6182 @rlbarter
I'm not overly familiar with the type of data that this package involves, but if you can't find any other reviewers, I'm happy to review in regards to basic completeness of documentation, etc. which if I recall, is mostly what JOSS review's involve anyway. My timeline would be similar to Yiqun's.
