### 2024-06-22 #6918 @arfon
Noting from the pre-review thread that both reviewers said they would need until the end of July to complete their reviews.

@markheckmann – while we're waiting, could I ask you to fix a couple of things please:

1. Please make sure there's a license file at the root of your repository with an OSI-approved license.
2. Could you also try and shorten your paper a bit if possible please? As a guide, we ask that JOSS papers are typically around 1000 words.

Thank you!
