+++ NEWS.md
@@ -1,5 +1,6 @@
-# OpenRepGrid 0.1.15
+# OpenRepGrid 0.1.15 (dev version)
 
+* add `LICENSE.md` file
 * `rgl` installation is now optional (close #41)
 * rename `ward` method to `ward.D` in `cluster` (fix #36) 
 * `indexDDI` and `indexUncertainty`, two dispersion indexes for dependency (e.g., situation-resource) grids added. Jon Raskin kindly helped with the documentation.
