+++ DESCRIPTION
@@ -17,7 +17,7 @@ Description: Analyze repertory grids, a qualitative-quantitative
     2004, ISBN: 978-0-470-09080-0). The package is part of the The package is part of the 
     <https://openrepgrid.org/> project.
 Version: 0.1.15
-Date: 2024-02-19
+Date: 2024-07-20
 Encoding: UTF-8
 URL: https://github.com/markheckmann/OpenRepGrid
 Imports:
@@ -69,7 +69,7 @@ Collate:
     'utils-import.r'
     'utils.r'
     'zzz.r'
-RoxygenNote: 7.3.0
+RoxygenNote: 7.3.1
 NeedsCompilation: no
 Suggests: 
     rgl,
