### 2024-07-20 da9a406
remove check_win_devel NOTES
