+++ vignettes/web/clustering.Rmd
@@ -13,7 +13,7 @@ vignette: >
 ```{r setup, echo=FALSE, message=FALSE, warning=FALSE}
 knitr::opts_chunk$set(collapse = TRUE, comment = "#")
 library(OpenRepGrid)
-settings(show.scale=FALSE, show.meta=FALSE)
+settings(show.scale = FALSE, show.meta = FALSE)
 ```
 
 ### Description
@@ -58,28 +58,28 @@ cluster(bell2010)
 The function also returns the reordered matrix invisibly. To see the reordered grid save it into a new object. To oppress the creation of a graphic set `print = FALSE`. 
 
 ```{r message=FALSE, warning=FALSE}
-x <- cluster(bell2010, print=FALSE)
+x <- cluster(bell2010, print = FALSE)
 x
 ```
 
 The function also allows to only cluster constructs or elements. To only cluster the constructs us the following code. Again a dendrogram is drawn and a grid with reordered constructs is returned.
 
 ```{r message=FALSE, warning=FALSE}
-x <- cluster(bell2010, along=1, print=FALSE)
+x <- cluster(bell2010, along = 1, print = FALSE)
 x
 ```
 
 To only cluster the elements set `along=2`.
 
 ```{r message=FALSE, warning=FALSE}
-x <- cluster(bell2010, along=2, print=FALSE)
+x <- cluster(bell2010, along = 2, print = FALSE)
 x
 ```
 
 To apply different distance measures and cluster methods us the arguments `dmethod` and `cmethod` (here `manhattan` distance and `single` linkage clustering).
 
 ```{r message=FALSE, warning=FALSE}
-x <- cluster(bell2010, dmethod="manh", cmethod="single", print=FALSE)
+x <- cluster(bell2010, dmethod = "manh", cmethod = "single", print = FALSE)
 x
 ```
 
@@ -87,22 +87,24 @@ To apply different methods to the constructs and the rows, use a two-step approa
 
 ```{r message=FALSE, warning=FALSE}
 # cluster constructs using default methods
-x <- cluster(bell2010, along=1, print=FALSE)       
+x <- cluster(bell2010, along = 1, print = FALSE)
 # cluster elements using manhattan distance and single linkage clustering
-x <- cluster(x, along=2, dm="manh", cm="single", print=FALSE)  
+x <- cluster(x, along = 2, dm = "manh", cm = "single", print = FALSE)
 x
 ```
 
 Some other options can be set. Paste the code into the R console to try it out. See `?cluster` for more information.
 
 ```{r eval=FALSE}
-cluster(bell2010, main="My cluster analysis")   # new title
-cluster(bell2010, type="t")                     # different drawing style
-cluster(bell2010, dmethod="manhattan")          # using manhattan metric
-cluster(bell2010, cmethod="single")             # do single linkage clustering
-cluster(bell2010, cex=1, lab.cex=1)             # change appearance
-cluster(bell2010, lab.cex=.7,                   # advanced appearance changes
-        edgePar = list(lty=1:2, col=2:1))
+cluster(bell2010, main = "My cluster analysis") # new title
+cluster(bell2010, type = "t") # different drawing style
+cluster(bell2010, dmethod = "manhattan") # using manhattan metric
+cluster(bell2010, cmethod = "single") # do single linkage clustering
+cluster(bell2010, cex = 1, lab.cex = 1) # change appearance
+cluster(bell2010,
+  lab.cex = .7, # advanced appearance changes
+  edgePar = list(lty = 1:2, col = 2:1)
+)
 ```
 
 #### Bootsrapped clustering
