+++ pkgdown.yml
@@ -25,7 +25,7 @@ articles:
   web/settings: web/settings.html
   web/visualization-bertin: web/visualization-bertin.html
   web/visualization-biplot: web/visualization-biplot.html
-last_built: 2024-07-20T11:35Z
+last_built: 2024-07-20T11:58Z
 urls:
   reference: https://docs.openrepgrid.org/reference
   article: https://docs.openrepgrid.org/articles
