### 2024-07-20 de4a89d
handle windbuilder NOTE
remove line 'VignetteBuilder: knitr' from DESCRIPTION. We do not
have vignettes. However, the .Rmd files for the pkgdown website
are stored in folder vignettes/web.
