+++ DESCRIPTION
@@ -80,4 +80,3 @@ Suggests:
     knitr,
     rmarkdown
 Roxygen: list(markdown = TRUE)
-VignetteBuilder: knitr
