### 2024-07-20 e78469f
Shorten paper, add docx, html to ignore
