+++ paper.Rmd
@@ -11,7 +11,7 @@ authors:
 affiliations:
 - name: DB InfraGO AG, Germany
   index: 1
-date: 01 January 2024
+date: 20 July 2024
 bibliography: paper.bib
 output:
   html_document:
@@ -37,40 +37,38 @@ library(OpenRepGrid)
 
 # Introduction
 
-The [OpenRepGrid R package](https://cran.r-project.org/web/packages/OpenRepGrid/index.html) is a software to analyze and visualize [repertory grid](https://en.wikipedia.org/wiki/Repertory_grid) (often abbreviated *grid* or *repgrid*) data. The software is open source and available on all major operating systems. The package is also the workhorse on which other packages of the [OpenRepGrid project](http://openrepgrid.org/), for example, [gridsampler](https://openresearchsoftware.metajnl.com/articles/10.5334/jors.150/) [@heckmann_gridsampler_2017] or [OpenRepGrid.ic](https://joss.theoj.org/papers/10.21105/joss.03292) [@heckmann_openrepgridic_2023], partially build upon. 
+[OpenRepGrid](https://cran.r-project.org/web/packages/OpenRepGrid/index.html) is an R package designed to analyze [repertory grid](https://en.wikipedia.org/wiki/Repertory_grid) (also *grid* or *repgrid*) data. The software is open source and available on all major operating systems. The package is also the workhorse on which other packages of the [OpenRepGrid project](http://openrepgrid.org/), for example, [gridsampler](https://openresearchsoftware.metajnl.com/articles/10.5334/jors.150/) [@heckmann_gridsampler_2017] or [OpenRepGrid.ic](https://joss.theoj.org/papers/10.21105/joss.03292) [@heckmann_openrepgridic_2023], partially build upon. 
 
 
 # Repertory Grid Technique 
 
-The repertory grid technique (RGT) is a data collection method which originated from *Personal Construct Theory (PCT)* [@kelly_psychology_1955]. It was originally designed as an instrument for psychotherapy to shed light on a client’s construction of the world. Over subsequent decades, the technique has been adopted in many other fields, including market, organizational, political, educational and sensory research [@fransella_manual_2004]. The data the RGT generates is *qualitative* and *quantitative*. On the qualitative side, the technique elicits the repertory of bipolar attributes (e.g. *smart vs. dull*, so called *constructs* in PCT terminology) an individual uses to make distinctions between entities of the world (e.g. different people, so called *elements* in PCT terminolgy). On the quatitative side, it requires rating each element on each elicited personal construct (e.g. *Martin* gets a score of 2 on the *quarrelsome = 1 vs. peaceful = 6* construct, indicating that Martin is quite quarrelsome). The result of the data collection procedure is a data matrix. The constructs are usually presented as matrix rows, the elements as columns and each cell contains the corresponding rating score. Figure 1 depicts a repertory grid data set, with the rows (constructs) and columns (elements) being clustered by similarity (see below for details). A thorough introduction to the repertory grid technique is given by @fransella_manual_2004.
[…]
