+++ NEWS.md
@@ -1,3 +1,5 @@
+# OpenRepGrid (development version)
+
 # OpenRepGrid 0.1.15
 
 * add `LICENSE.md` file
