### 2024-07-20 #6918 @markheckmann
@arfon , thanks for moving the review forward.

1. The repo now contains a `LICENSE.md` file with a GPL license as well as a reference to it in the `DESCRIPTION` file.
2. I shortened the paper. The word count (exluding references) is now ~1,400. I tried to still keep it informative for readers who may not be familiar with the repertory grid technique. I would be grateful if this length was deemed acceptable.
3. Version `0.1.15` was published on [CRAN](https://cran.r-project.org/web/packages/OpenRepGrid/index.html) today.

@yiqunchen, @rlbarter: Thanks for agreeing to review this paper.
