### 2024-07-26 #6918 @rlbarter
Yes, I plan to spend some time on this today and early next week.
