### 2024-07-26 markheckmann/OpenRepGrid#43 @rlbarter
JOSS Review: add community guidelines
G'day Mark!

I'm working on my review of OpenRepGrid for JOSS (https://github.com/openjournals/joss-reviews/issues/6918), and everything looks good in my checklist except for "**Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support?"

If you could add some community guidelines when you have a chance, and let me know, and I'll be able to conclude my review.
