### 2024-07-30 5f58144
Add link to contributing guidelines (#43)
