### 2024-07-30 markheckmann/OpenRepGrid#43 @markheckmann
Dear @rlbarter ,
thanks for your review and the issue. A contributing guideline exists under the package [docs page](https://docs.openrepgrid.org/CONTRIBUTING.html). As it might be hard to find it, I added a new [section to the README.md ](https://github.com/markheckmann/OpenRepGrid?tab=readme-ov-file#contributing-to-openrepgrid) which links to the page. Please let me know, if that works for you.
