+++ dev/pkgdown.yml
@@ -25,7 +25,7 @@ articles:
   web/settings: web/settings.html
   web/visualization-bertin: web/visualization-bertin.html
   web/visualization-biplot: web/visualization-biplot.html
-last_built: 2024-07-30T17:40Z
+last_built: 2024-07-30T17:49Z
 urls:
   reference: https://docs.openrepgrid.org/reference
   article: https://docs.openrepgrid.org/articles
