+++ README.Rmd
@@ -60,6 +60,10 @@ And for a first impression of the package features type
 
     demo(OpenRepGrid)
 
+## Documentation
+
+Visit the package documentation website: <https://docs.openrepgrid.org/>
+
 
 ## Contributing to OpenRepGrid
 
