### 2024-09-13 #6918 @yiqunchen
I have completed my review and the paper looks great! The only request I have for the paper/repo is that it would be great to include an example set of raw input so users could understand the data structure of repertory grid data better. Thank you for a great package and apologies for the dealy on my end again.
