### 2024-09-14 c9750e9
tweak data import docs for JOSS review
JOSS Review: <https://github.com/openjournals/joss-reviews/issues/6918#issuecomment-2350417777>
