### 2024-09-14 #6918 @markheckmann
HI @yiqunchen , thanks for completing the review and your suggestion.

I included two new sample `.txt` files in the package ([commit](https://github.com/markheckmann/OpenRepGrid/commit/c9750e9024c0823b75c14dc1781c94cbbfce4342)). They will be published with the next CRAM release. Two Excel files were already included. The [documentation ](https://docs.openrepgrid.org/index.html) (source on [gh-pages branch](https://github.com/markheckmann/OpenRepGrid/tree/gh-pages) ) also contains several places where the different file input formats are documented (see below).

**Would that suffice for you?**

----

1. Overview on loading data: <https://docs.openrepgrid.org/articles/web/loading.html#loading-openrepgrid-files>

2. Details for each input type:
    * `.txt`: <https://docs.openrepgrid.org/reference/importTxt.html> (will be auto-updated with next CRAN release)
    * `.xlsx`: <https://docs.openrepgrid.org/reference/importExcel.html>
    * ...
