### 2024-12-01 #6918 @arfon
As it seems like @yiqunchen is currently not available, I will make the decision that [this response from the author](https://github.com/openjournals/joss-reviews/issues/6918#issuecomment-2351000853) satisfies the feedback from the reviewer.

@markheckmann – looks like we're very close to being done here. I will circle back here later this week, but in the meantime, please give your own paper a final read to check for any potential typos etc.
