### 2024-12-02 5734880
Bump version to 0.1.16, update NEWS
