+++ NEWS.md
@@ -1,4 +1,6 @@
-# OpenRepGrid (development version)
+# OpenRepGrid 0.1.16
+
+* Improved documentation as required during [JOSS](https://github.com/openjournals/joss-reviews/issues/6918) review.
 
 # OpenRepGrid 0.1.15
 
