### 2024-12-02 c9e3bcb
fix typos, improve grammar
