### 2024-12-02 7c2c3aa
fix more typos and improve language
