+++ paper.md
@@ -38,7 +38,7 @@ pre {
 
 # Repertory Grid Technique 
 
-The repertory grid technique (RGT) is a data collection method that originated from *Personal Construct Theory (PCT)* [@kelly_psychology_1955]. It was originally designed as an instrument for psychotherapy to shed light on a client’s construction of the world. Over subsequent decades, the technique has been adopted in many other fields, including market, organizational, political, educational or sensory research [@fransella_manual_2004]. The data the RGT generates is *qualitative* and *quantitative*. On the qualitative side, the technique elicits the repertory of bipolar attributes (e.g. *smart vs. dull*, so-called *constructs* in PCT terminology) that an individual uses to make distinctions between entities of the world (e.g. different people, referred to as *elements* in PCT terminolgy). On the quantitative side, it requires rating each element on each elicited personal construct (e.g. *Martin* has a score of 2 on the *quarrelsome = 1 vs. peaceful = 6* construct, indicating that Martin is rather quarrelsome). The result of the data collection procedure is a matrix. The constructs are presented as rows, the elements as columns, and each cell contains the corresponding rating score. Figure 1 depicts a grid data set, with the rows (constructs) and columns (elements)  clustered by similarity (see below for details). A thorough introduction to the repertory grid technique is given by @fransella_manual_2004.
+The repertory grid technique (RGT) is a data collection method that originated from *Personal Construct Theory (PCT)* [@kelly_psychology_1955]. It was originally designed as an instrument for psychotherapy to shed light on a client’s construction of the world. Over subsequent decades, the technique has been adopted in many other fields, including market, organizational, political, educational or sensory research [@fransella_manual_2004]. The data the RGT generates is *qualitative* and *quantitative*. On the qualitative side, the technique elicits the repertory of bipolar attributes (e.g. *smart vs. dull*, so-called *constructs* in PCT terminology) that an individual uses to make distinctions between entities of the world (e.g. different people, referred to as *elements* in PCT terminology). On the quantitative side, it requires rating each element on each elicited personal construct (e.g. *Martin* has a score of 2 on the *quarrelsome = 1 vs. peaceful = 6* construct, indicating that Martin is rather quarrelsome). The result of the data collection procedure is a matrix. The constructs are presented as rows, the elements as columns, and each cell contains the corresponding rating score. Figure 1 depicts a grid data set, with the rows (constructs) and columns (elements)  clustered by similarity (see below for details). A thorough introduction to the repertory grid technique is given by @fransella_manual_2004.
 
 ![**Figure 1.** Example of a repertory grid dataset (with rows and columns clustered by similarity).](img/01-bertin-clustered.png)
 
@@ -47,7 +47,7 @@ The repertory grid technique (RGT) is a data collection method that originated f
 
 While it is possible to work with repertory grids directly without further processing, it is common to submit grid data to statistical analysis [e.g. @fransella_manual_2004]. For this purpose, software packages have been developed since the 1960s [@sewell_computerized_1992], including Enquire Within [@mayes_enquire_2008], GridStat [@bell_gridstat_2009], GridCor [@feixas_gridcor:_2002], Idiogrid [@grice_idiogrid:_2002], Rep 5 [@gaines_rep_2009], GridSuite [@fromm_gridsuite_2006], rep:grid [@rosenberger_vademecum_2015]. Despite the numerous software packages available, several issues are common among them: 
 
-* All listed softwares programs are closed source or at least not available in a public repository.
+* All listed software programs are closed source or at least not available in a public repository.
 * No software offers all analysis methods devised in the literature. 
 * None of the programs can be extended by the user, for example, by adding or modifying features. 
[…]
