### 2024-12-02 f99f829
CRAN submission v0.1.16
