+++ cran-comments.md
@@ -1,12 +1,8 @@
-## Submission info
-
-Fix v0.1.14 CRAN check NOTES: <https://cran.r-project.org/web/checks/check_results_OpenRepGrid.html>
-
 ## Test environments
 
-* MacOS 14.5, R 4.4.1
-* Microsoft Windows Server 2022, R 4.4.1
-* Ubuntu 22.04.4 LTS, R 4.4.1
+* MacOS 14.7.1, R 4.4.2
+* Microsoft Windows Server 2022, R 4.4.2
+* Ubuntu 22.04.5 LTS, R 4.4.2
 
 ## R CMD check results
 
@@ -15,4 +11,3 @@ Fix v0.1.14 CRAN check NOTES: <https://cran.r-project.org/web/checks/check_resul
 ## Reverse dependencies
 
 There are no reverse dependencies.
-
