+++ NEWS.md
@@ -1,3 +1,7 @@
+# OpenRepGrid (dev version)
+
+* ...
+
 # OpenRepGrid 0.1.16
 
 * Improved documentation as required during [JOSS](https://github.com/openjournals/joss-reviews/issues/6918) review.
