### 2024-12-02 #6918 @markheckmann
Dear @arfon ,

1. I proofread the paper again and removed a few typos. It should be fine now.
2. A new R package version (0.1.16) containing the changes requested by the reviewers was published on [CRAN](https://cran.r-project.org/web/packages/OpenRepGrid/index.html) today
3. The package code was uploaded to Zenodo: https://zenodo.org/records/14262588 DOI: [10.5281/zenodo.14262588](https://doi.org/10.5281/zenodo.14262588)
