### 2024-12-07 #6918 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
