### 2024-12-18 #6918 @arfon
@markheckmann – I spotted a minor issue with the figure captions. Could you please merge this PR? https://github.com/markheckmann/OpenRepGrid/pull/44
