### 2024-12-19 #6918 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
